+++ CHANGELOG.md
@@ -5,7 +5,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Update the APIs of the following add-ons:
+  - Automation Framework version 0.15.0;
+  - Selenium version 15.8.0;
 
+### Fixed 
+- Add a graphql object to ClientApi.
 
 ## [1.10.0] - 2021-10-11
 ### Added